+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   assert expectations
 - refactor code that sets 'TMUX_PLUGIN_MANAGER_PATH' global env var
 - stop using global variable for 'tpm path'
+- support defining plugins via `set -g @plugin` in sourced files as well
 
 ### v2.0.0, 2015-07-07
 - enable overriding default key bindings